+++ board_files/include/dispatch/general_dispatch.php
@@ -4,8 +4,6 @@ if (!defined('NELLIEL_VERSION'))
     die("NOPE.AVI");
 }
 
-require_once INCLUDE_PATH . 'thread_functions.php';
-
 function nel_general_dispatch()
 {
     $module = (isset($_GET['module'])) ? $_GET['module'] : null;