+++ board_files/include/output/management/file_filter_panel.php
@@ -10,8 +10,8 @@ function nel_render_file_filter_panel()
     $render = new NellielTemplates\RenderCore();
     $render->startRenderTimer();
     $render->getTemplateInstance()->setTemplatePath(TEMPLATE_PATH);
-    nel_render_general_header($render, null, null, array('header' => 'MANAGE_BOARD',
-        'sub_header' => 'MANAGE_BANS'));
+    nel_render_general_header($render, null, null, array('header' => 'Board Management',
+        'sub_header' => 'Bans'));
     $dom = $render->newDOMDocument();
     $render->loadTemplateFromFile($dom, 'management/file_filter_panel.html');
     $dom->getElementById('add-file-filter-form')->extSetAttribute('action', PHP_SELF .