+++ board_files/include/output/management/login_page.php
@@ -9,7 +9,7 @@ function nel_render_login_page()
     $render = new NellielTemplates\RenderCore();
     $render->startRenderTimer();
     $render->getTemplateInstance()->setTemplatePath(TEMPLATE_PATH);
-    nel_render_general_header($render, null, null, array('header' => 'MANAGE_LOGIN'));
+    nel_render_general_header($render, null, null, array('header' => 'Management Login'));
     $dom = $render->newDOMDocument();
     $render->loadTemplateFromFile($dom, 'management/login.html');
     $dom->getElementById('login-form')->extSetAttribute('action', PHP_SELF . '?manage=login');