+++ board_files/include/output/management/thread_panel.php
@@ -11,7 +11,7 @@ function nel_render_thread_panel_main($board_id)
     $render = new NellielTemplates\RenderCore();
     $render->startRenderTimer();
     $render->getTemplateInstance()->setTemplatePath(TEMPLATE_PATH);
-    nel_render_general_header($render, null, $board_id, array('header' => 'MANAGE_BOARD', 'sub_header' => 'MANAGE_THREADS'));
+    nel_render_general_header($render, null, $board_id, array('header' => 'Board Management', 'sub_header' => 'Threads'));
     $dom = $render->newDOMDocument();
     $render->loadTemplateFromFile($dom, 'management/thread_panel.html');
     $dom->getElementById('board_id_field')->extSetAttribute('value', $board_id);
@@ -35,7 +35,7 @@ function nel_render_thread_panel_main($board_id)
         unset($result);
 
         $expand_thread_button = $temp_thread_row->getElementById('expand-thread-button-');
-        $expand_thread_button->extSetAttribute('value', nel_stext('FORM_EXPAND') . ' ' . $thread['thread_id']);
+        $expand_thread_button->extSetAttribute('value', _gettext('Expand') . ' ' . $thread['thread_id']);
         $expand_thread_button->changeId('expand-thread-button-' . $thread['thread_id']);
         $thread_post_number = $temp_thread_row->getElementById('thread-post-number-');
         $thread_post_number->setContent($thread['thread_id']);
@@ -119,7 +119,7 @@ function nel_render_thread_panel_expand($board_id, $thread_id)
     $render = new NellielTemplates\RenderCore();
     $render->startRenderTimer();
     $render->getTemplateInstance()->setTemplatePath(TEMPLATE_PATH);
-    nel_render_general_header($render, null, $board_id, array('header' => 'MANAGE_BOARD', 'sub_header' => 'MANAGE_THREADS'));
+    nel_render_general_header($render, null, $board_id, array('header' => 'Board Management', 'sub_header' => 'Threads'));
     $dom = $render->newDOMDocument();
     $render->loadTemplateFromFile($dom, 'management/thread_panel_expand.html');
     $dom->getElementById('thread-list-form')->extSetAttribute('action', PHP_SELF .