+++ imgboard.php
@@ -1,5 +1,7 @@
 <?php
-define('NELLIEL_VERSION', 'v0.9.6.7'); // Version
+define('NELLIEL_VERSION', 'v0.9.6.9'); // Version
+define('NELLIEL_COPYRIGHT', '2010-2018 Nelliel Project'); // Copyright line
+define('NELLIEL_PACKAGE', 'Nelliel'); // Package
 define('BASE_PATH', realpath('./') . '/'); // Base path for script
 define('FILES_PATH', BASE_PATH . 'board_files/'); // Base board files path
 define('INCLUDE_PATH', FILES_PATH . 'include/'); // Base include files path
@@ -7,7 +9,8 @@ define('CONFIG_PATH', BASE_PATH . 'configuration/'); // Base cache path
 define('LIBRARY_PATH', FILES_PATH . 'libraries/'); // Libraries path
 define('PLUGINS_PATH', FILES_PATH . 'plugins/'); // Base plugins path
 define('TEMPLATE_PATH', FILES_PATH . 'templates/nelliel/'); // Base template path
-define('LANGUAGE_PATH', FILES_PATH . 'languages/'); // Language files path
+define('LANGUAGE_PATH', FILES_PATH . 'languages/'); // Language path
+define('LOCALE_PATH', LANGUAGE_PATH . 'locale/'); // Locale files path
 define('CACHE_PATH', FILES_PATH . 'cache/'); // Base cache path
 define('WEB_PATH', BASE_PATH . 'web/'); // Base cache path
 define('SQLITE_DB_DEFAULT_PATH', FILES_PATH); // Base SQLite DB location
@@ -17,10 +20,14 @@ require_once LIBRARY_PATH . 'phpDOMExtend/autoload.php';
 require_once LIBRARY_PATH . 'NellielTemplates/autoload.php';
 require_once LIBRARY_PATH . 'portable-utf8/portable-utf8.php';
 require_once LIBRARY_PATH . 'password_compat/lib/password.php';
+require_once LIBRARY_PATH . 'SmallPHPGettext/autoload.php';
 require_once INCLUDE_PATH . 'initializations.php';
 require_once INCLUDE_PATH . 'database.php';
 require_once INCLUDE_PATH . 'general_functions.php';
 require_once INCLUDE_PATH . 'language/language.php';
+
+nel_load_language_library(LOCALE_PATH . DEFAULT_LOCALE . '/LC_MESSAGES/nelliel.po');
+
 require_once INCLUDE_PATH . 'derp.php';
 require_once INCLUDE_PATH . 'crypt.php';